fix run modisco script to run sherlock

+++ sherlock/modisco/submit_modisco_jobs.sh
@@ -4,17 +4,19 @@ experiment=$1
 round=$2
 oak_dir=$3
 
-cd /scratch/users/vir/; rm -rf my_scripts; mkdir my_scripts
-cd /scratch/users/vir/my_scripts
-git clone --depth 1 --branch main https://github.com/viramalingam/tf-atlas-pipeline.git
-chmod -R 777 tf-atlas-pipeline
-cd tf-atlas-pipeline/sherlock/modisco/
+if ![[ -f /scratch/users/vir/my_scripts ]]; then
+    cd /scratch/users/vir/; rm -rf my_scripts; mkdir my_scripts
+    cd /scratch/users/vir/my_scripts
+    git clone --depth 1 --branch main https://github.com/viramalingam/tf-atlas-pipeline.git
+    chmod -R 777 tf-atlas-pipeline
+fi
+cd /scratch/users/vir/my_scripts/tf-atlas-pipeline/sherlock/modisco/
 
 if [ -f $oak_dir/shap/$round/$experiment/profile_scores.h5 ] && [ -f $oak_dir/shap/$round/$experiment/counts_scores.h5 ] ; then
     if [ -f $oak_dir/modisco/$round/$experiment/profile/modisco_results.h5 ] && [ -f $oak_dir/modisco/$round/$experiment/counts/modisco_results.h5 ] ; then
         echo "modisco run exists"
     else
-        if ![[-f $oak_dir/modisco/$round/$experiment]]; then
+        if ![[ -f $oak_dir/modisco/$round/$experiment ]]; then
             mkdir $oak_dir/modisco/$round/$experiment
         fi
         modisco_dir=$oak_dir/modisco/$round/$experiment
